requirements.txt
asgiref==3.6.0
autopep8==2.0.1
colorama==0.4.6
Django==4.1.6
django-cors-headers==3.13.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
nose==1.3.7
pathlib==1.0.1
phonenumbers==8.13.6
phonenumberslite==8.13.6
Pillow==9.4.0
pycodestyle==2.10.0
pymongo==4.3.3
pypng==0.20220715.0
pytz==2022.7.1
qrcode==7.4.2
sqlparse==0.2.4
typing_extensions==4.4.0
tzdata==2022.7
gunicorn==20.0.4